Fix formatting and typos in Share
