+++ src/Idfy.SDK/Services/Share/Entities/Authentication.cs
@@ -3,17 +3,21 @@
     public class Authentication
     {
         /// <summary>
-        ///The identityprovider type (Norwegian BanKID, SwedishBankID, Nemid, etc)
+        ///The identity provider type (Norwegian BanKID, SwedishBankID, Nemid, etc)
         /// </summary>
         public IdentityProviderType IdentityProvider { get; set; }
+        
         /// <summary>
-        /// The uniqueID from the e-ID, this ID is unique for the user and is the same every time the user logs on. This is not a sensitiv ID and you could store this to identify the user in you database. Remark: The Swedish BankID do not have an unique ID.
+        /// The uniqueID from the e-ID, this ID is unique for the user and is the same every time the user logs on. This is not a sensitive ID
+        /// and you could store this to identify the user in you database. Remark: The Swedish BankID do not have an unique ID.
         /// </summary>
         public string IdentityProviderUniqueId { get; set; }
+        
         /// <summary>
         /// Social security number 
         /// </summary>
         public string Ssn { get; set; }
+        
         /// <summary>
         /// The users date of birth format ddMMyy
         /// </summary>
