+++ src/Idfy.SDK/Services/Share/Entities/ContactDetails.cs
@@ -6,14 +6,17 @@
         /// Name to present to the recipient
         /// </summary>
         public string Name { get; set; }
+        
         /// <summary>
-        /// Phonenumber recipient can contact 
+        /// Phone number recipient can contact 
         /// </summary>
         public string Phone { get; set; }
+        
         /// <summary>
         /// Email recipient can contact 
         /// </summary>
         public string Email { get; set; }
+        
         /// <summary>
         /// Web page the recipient can visit
         /// </summary>
