+++ src/Idfy.SDK/Services/Share/Entities/Notification.cs
@@ -6,9 +6,9 @@
     public class Notification
     {
         public NotificationSetup Setup { get; set; }
+        
         /// <summary>
-        /// Language for the notifications
-        /// Defaults to Norwegian
+        /// Language for the notifications. Defaults to Norwegian
         /// </summary>
         public Language Language { get; set; }
     }
