+++ src/Idfy.SDK/Services/Share/Entities/ServerUpload.cs
@@ -1,0 +1,13 @@
+namespace Idfy.Share.Entities
+{
+    public class ServerUpload
+    {
+        public string Id { get; set; }
+        
+        public string Name { get; set; }
+        
+        public string Content { get; set; }
+        
+        public UploadType Type { get; set; }
+    }
+}
