+++ src/Idfy.SDK/Services/Share/Entities/ShareCreateOptions.cs
@@ -15,11 +15,13 @@ namespace Idfy.Share.Entities
         /// </summary>
 
         public IEnumerable<Recipient> Recipients { get; set; }
+        
         public ContactDetails ContactDetails{ get; set; }
         /// <summary>
         /// List of people sending this share.
         /// </summary>
         public IEnumerable<Sender> Senders { get; set; }
+        
         /// <summary>
         /// Optional settings for advanced configuration
         /// </summary>
