+++ src/Idfy.SDK/Services/Share/Entities/ShareUpdateOptions.cs
@@ -6,10 +6,12 @@
         /// Title of the secure share, will be presented to the recipient 
         /// </summary>
         public string Title { get; set; }
+        
         /// <summary>
-        /// Descripition of secure share, will be present in standard notificaitons
+        /// Description of secure share, will be present in standard notifications
         /// </summary>
         public string Description { get; set; }
+        
         public ContactDetails ContactDetails { get; set; }
     }
 }
