+++ src/Idfy.SDK/Services/Share/Entities/Upload.cs
@@ -6,10 +6,12 @@
         /// Unique id for share 
         /// </summary>
         public string Id { get; set; }
+        
         /// <summary>
         /// Name of upload, will be presented to the user
         /// </summary>
         public string Name { get; set; }
+        
         /// <summary>
         /// What kind of content is it. Defaults to file
         /// </summary>
