+++ src/Idfy.SDK/Services/Share/Entities/UploadResponse.cs
@@ -1,0 +1,25 @@
+using System;
+
+namespace Idfy.Share.Entities
+{
+    /// <summary>
+    /// Upload information
+    /// </summary>
+    public class UploadResponse
+    {
+        /// <summary>
+        /// Name of content
+        /// </summary>
+        public string Name { get; set; }
+        
+        /// <summary>
+        /// If the recipient has opened the file 
+        /// </summary>
+        public bool Opened { get; set; }
+        
+        /// <summary>
+        /// When the file was last opened
+        /// </summary>
+        public DateTime OpenedDate { get; set; }
+    }
+}
